First Rendition
Still a few errors to work with. Like 0 index doesnt work with the buttons so there has to be a blank button at the top. We could probably skip this button so it doesnt show up. Everything else works good. Next step is to make it take stuff from a database instead of a json file